Update STRONG_PASSWORD
Made sure the prompt was properly commented out, 

I also want to provide some explanation for my solution. 

I set up boolean variables for each of the necessary char types(upper, lower, number, and special characters)
I also copied the string containing all of the special characters from the prompt, this is because for all of the other types of chars we have sequential ascii values that can be used for comparison, but for the special characters that is not the case

the outer loop iterates through the entered password, at each point it checks the type of character(lower, upper, number, or special), I initially had it checking the associated boolean for each of these in the same if/if else conditional statement
--the problem with that approach was that once the lower, upper and number flags were set to false every subsequent char entered the nested loop, so to reduce the number of unnecessary loops I changed the conditional statements to assess the type of char before checking the flags

my approach uses the fact that boolean values in C++ are either 1(true) or 0(false), so after checking every char in the password string we can sum the boolean values to find the missing types of chars.

the final thing to check was the length of the password. If the password's length was too short we could ignore the flags entirely so I checked if the length of the password was less than the required additional chars from the boolean flags.